add "fake-buildid" USE flag on firefox
When firefox is running and portage updates it, firefox desperately want
user to restart it without any way to disable this.

IMHO, it is really annoying as it restores all tabs but not their
states. Moreover, if you are running a multi-user system, you will
always have someone running it.

When this USE flag is set, firefox build ID will never changed which
prevent firefox "upgrade" detection.

Note: if running firefox version and installed firefox version are too
different, tab may crash... or not, who knows...